EDT-57 Added DV_IDENTIFIER to archetype editor, read and write to ADL (tested) and XML (not tested). Also have much of the work for DV_CURRENCY done - awaiting Tom's modelling on this.

+++ DataConstraints/CountConstraint.vb
@@ -17,9 +17,9 @@ Option Strict On
 Public Class Constraint_Count
     Inherits Constraint_with_value
 
-    Protected mMinVal As Single = 0
-    Protected mMaxVal As Single = 0
-    Protected mAssumedValue As Single
+    Protected mMinVal As Double = 0
+    Protected mMaxVal As Double = 0
+    Protected mAssumedValue As Double
     Protected mHasMaxVal As Boolean
     Protected mHasMinVal As Boolean
     Protected mIncludeMax As Boolean = True
